feat: use positional range arg instead of from/to
This drops the from/to args in favor for a positional
range argument. This makes for more convenient usage.
Instead of:

`gtag --from sha1 --to sha2 --pattern tag-##1`

We can now write:

`gtag sha1..sha2 --pattern tag-##1`

Please note that commit ranges are always *inclusive*
which is in contrast to how Git normaly handles commit
ranges. This was a choise made for convenience.